fix: Implementación de Blueprints
