+++ blueprints/compras/routes.py
@@ -1,0 +1,5 @@
+from . import compras_bp
+
+@compras_bp.route('/')  
+def compras():
+    return 'Módulo de compras'
