+++ blueprints/inventario/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+inventario_bp = Blueprint('inventario', __name__, url_prefix='/inventario', template_folder='templates')
+
+from . import routes
