+++ blueprints/inventario/routes.py
@@ -1,0 +1,5 @@
+from . import inventario_bp
+
+@inventario_bp.route('/')
+def inventario():
+    return 'Módulo de inventario'
