+++ blueprints/inventario/templates/inventario.html
