+++ blueprints/pedidos/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+pedidos_bp = Blueprint('pedidos', __name__, url_prefix='/pedidos', template_folder='templates')
+
+from . import routes
