+++ blueprints/pedidos/routes.py
@@ -1,0 +1,5 @@
+from . import pedidos_bp
+
+@pedidos_bp.route('/')
+def pedidos():
+    return 'Módulo de pedidos'
