+++ blueprints/produccion/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+produccion_bp = Blueprint('produccion', __name__, url_prefix='/produccion', template_folder='templates')
+
+from . import routes
