+++ blueprints/produccion/routes.py
@@ -1,0 +1,5 @@
+from .import produccion_bp
+
+@produccion_bp.route('/')
+def produccion():
+    return 'Módulo de producción'
