+++ blueprints/proveedores/routes.py
@@ -1,0 +1,5 @@
+from . import proveedores_bp
+
+@proveedores_bp.route('/')
+def proveedores():
+    return 'Módulo de proveedores'
