+++ blueprints/recetas/routes.py
@@ -1,0 +1,5 @@
+from . import recetas_bp
+
+@recetas_bp.route('/')
+def recetas():
+    return 'Módulo de recetas'
