+++ blueprints/usuarios/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+usuarios_bp = Blueprint('usuarios', __name__, url_prefix='/usuarios', template_folder='templates')
+
+from . import routes
