+++ blueprints/usuarios/routes.py
@@ -1,0 +1,5 @@
+from . import usuarios_bp
+
+@usuarios_bp.route('/')
+def usuarios():
+    return 'Módulo de usuarios'
