+++ blueprints/ventas/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+ventas_bp = Blueprint('venta', __name__, url_prefix='/ventas', template_folder='templates')
+
+from . import routes
