+++ blueprints/ventas/templates/paquetes.html
